First iteration of the "Entity" class system
This is how the Entity class works:

5 components: stats, stateMachine, AI, physical, visual

Stats - Contains important numbers like health and movement speed. Also kills the entity when health is 0. This component should be manually attached to the root object of the entity so we can tweak the entity's base stats per-prefab in the inspector.

StateMachine - iterates over current EntityStates and runs the methods in them if they are active. EntityStates can be seen as "behaviour" scripts that truly define what this enemy does. The Entity class system is only a basic blueprint. Adding functionalities happens in the EntityStates.

AI - Currently just holds a list of struct instances (target's Entity class instance, target GameObject instance, and target's EntityStats). The player should really be the only target as of now.

Physical - Holds rigidbody and collider references and defines basic movements. Use this component to move the entity. LMK if you think of other universally useful movement functions. IF IT IS SPECIFIC TO AN ENEMY TYPE, WRITE IT AS A SCRIPT THAT INHERITS EntityState.

Visual - Holds any sprite renderers for the entity and its animator. Not much going on here rn.

Basic structure of an enemy prefab:
<components>

-root object <Entity, EntityStats, Rigidbody2D, CircleCollider2D>
  -Visual <Animator, SpriteRenderer> (the name should be "Visual" as the EntityVisual script GameObject.Finds it, we can change this later if complications occur)

How to define a new behaviour:

Assets > Scripts > EntityClass > EntityStates > TestMove.cs
This is an example of how EntityStates are used to define behaviours for entities.
Basic outline of EntityStates are as follows:

EntityState inherits ScriptableObject. Entity class component has a list of EntityStates exposed in the inspector, meaning you can adjust which behaviours/states the prefab will have on spawning and whether those states are active or not. To give an entity an existing state, drag in an EntityState-inherited ScriptableObject asset in Assets > Scripts > EntityClass > EntityStates > ScriptableObjects folder. To make a new behaviour/state, go to Assets > Scripts > EntityClass > EntityStates and create a new .cs file following the example of TestMove.cs. Then, in Assets > Scripts > EntityClass > EntityStates > ScriptableObjects, right click, Create > EntityState > select the state name you just created. This asset can now be dragged into the list of EntityStates in the Entity class component of the root object by anyone.